+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.13.3 (28.6.2014)
 
 - support for `s/Uuid` via latest `ring-swagger`.
+- fail-fast (with client-typos): remove default implementation from `compojure.api.meta/restructure-param`
 
 ## 0.13.2 (28.6.2014)